remove brittle and not particularly useful test
If the test is run while monit/munin connects to the AMI, then the
test breaks.

Also, it was more useful when "sccp reload" was unloading the whole
module then loading it again. Currently, the module reload only reload
the configuration, and it doesn't close/create new file descriptor
anyway.