remove faulty LDAP search scenarios
reason: we cant control the ldap from xivo-acceptance. There are still
scenarios about faulty LDAP server for searches from phones, but the
current implementation is sufficient. The current implementation is
not suited to CTI searches, because the CTI is not signaled when the
LDAP server configuration is changed, and even if it was, it would not
load the new config.

Also, this kind of tests will be executed automatically by the tests
for xivo-dird.