add unnecessary click to have the mouse avoid an element
reason: given I have a filtered list of users. When removing the first
user in the list via the Delete button, the mouse of Selenium may
hover on the blue arrow button top right of the screen, then go down
and click on the Delete of the dropdown, instead of the Delete with
the cross.