use absurdly high ID to workaround test precondition
Currently, the precondition for these 2 tests doesn't work for these
scenarios if there's already a user/line with the given ID, so use an ID
which has little chance of being already in use to workaround the bug.

Problem is that the xivo-ctid's cache is not updated correctly when a
user/line is removed by these steps.