restart browser before every test
This is an attempt to fix a problem with failing tests.
When tests run on jenkins there are 4 that fail.
But, these tests pass when run one by one.

My current theory is that these tests fail because some javascript
isn't running anymore. Maybe the browser can't fetch the js files,
or maybe the browser has been running for so long and a bug
is triggered by another test, thus freezing the javascript.

To test this theory, we're restarting the browser before each step
so that the javascript is loaded in a clean state.