create new session before each request
To prevent the connection from being pooled/reused, which is sometimes
causing "ConnectionError" against xivo-confd, because when requests test if
a pooled connection is closed or not, if it's open then xivo-confd might
still closes it between the time of the check and the time requests
use it, leading to random ConnectionError. There's probably other way
to fix this problem, but meh.