fix scenario "Restart" of asterisk.feature
There was a sort of race condition which was making the test failing:

* the step "When I stop "asterisk"" was executed and asterisk was stopped
* just after that, monit was seeing that asterisk (or xivo-ctid) was not
  up and was starting it
* then the step "Then the service "asterisk" is no longer running" was
  executed, but was failing since asterisk was actually restarted by
  monit between the execution of the two steps

So to make the test more reliable, we stop monit before checking that
the service are stopped, and only start it after these checks.

Also remove the step "When I wait for the service "asterisk" to stop"
since stopping a service with systemctl already waits that the process
exit before returning.