set the tenant_uuid on the auth and confd client
this will most likely break some tests but it is more in line with
the desired state than the current implementation.

Without this change all queries which do not specify a token will use the
master tenant uuid which is not the tenant of any entity. This happens
because we are using a xivo_service authentification which in the master
tenant.