attended transfer: use linphone instead of chan_test
reason: with chan_test we cannot know if the channel is on hold or not.
Also, chan_test is not something we necessary want to keep or to evolve.
It has been created before using linphone/real softphone.
The test should be easily if the incoming call have the same
properties/functionality as the other calls