user import: test the wazo-auth "rollback" on import error
If wazo-auth returns an error during the creation of a user the other valid
users should not be left in the DB without a valid matching ipbx user.

Since a user with username "john" is imported and "john" already exist the other
user "one" should not be created in wazo-auth.

This will allow the admin to fix its CSV and retry or create "one" manually with
getting a 409 because "one" already exist