transfer: wait for call processing
reason: fix race condition
when User A make a call, its phone is automatically in talking mode
(even if the call is not bridged). Incall is to talking because is in
a parking lot. So when doing an attended transfer, User A may not be
bridged with Incall, and the API will not be able to make a attended
transfer