user: avoid race condition with register
why:
When line is associated to user, chatd will receive this event and
create this line in its database. Before that event, all
DeviceStateChange triggered by asterisk will be ignored.

In this step, phone can register to asterisk BEFORE that chatd receive
the user_line_associated event, then it will ignore this register. The
result will be a desync of the user status and a warning in log:

No event received for chatd_presence_updated

Waiting for chatd_presence_udpated will ensure that user and line are
created in wazo-chatd and ready to update state properly