push: test that a cancel event is published when the caller hangs up
the event is used to stop the mobile from ringing.

When the caller hangs up before the mobile phone gets REGISTERed on Asterisk
then the mobile won't receive it's INVITE and it's BYE. In that case, there's
nothing to notify the mobile that the call is no longer waiting to be answered.