push: remove an invalid push
This scenario reject the call from the plastic phone of the user. This does not
stop the push dial because we are still waiting for the user to answer its
mobile.

If we wanted to stop the ringing when the plastic phone rejects the call that
stasis app waiting for mobile would have to be modified to stop waiting and
return. This change would make this test valid but it is not the current
behavior.