user: randomize line names
Why:

* Running two tests in a row with the same line name can cause
unexpected disconnections of SIP client in the second test
* Technically, Asterisk deregisters the AOR of the first test because
the qualify expired in the second test. Because the two AORs in each
test have the same name, IP address and port, there is a qui pro quo.
This may be a bug in Asterisk, but it's such a corner case that it's not
worth investigating.