asterisk queuelog: make EXITWITHTIMEOUT more reliable
Queue timeouts are hard to predict. The queue has a timeout that will hangup a
call when this timeout is reached. But there's also a timeout on call
distribution. If a call is being distributed to an agent the call will NOT
timeout until the agent timeout happens.

It is possible to spend more time than TIMEOUT in a queue before the EXITTIMEOUT
depending on the time at which you enter a call distribution loop.

In this scenario the agent is being distributed the first call then after 14
seconds the caller hangs up which is unnecessary because the queue already
dropped this call.

When the second call enters the queue the distribution loop could send call 1 OR
call 2 to the agent and this is where it gets unpredictable.

I changed the test to leave the call in the queue until the timeout happens
instead of trying to hang up once the timeout has been reached and generating an
abandonned call every once in a while