Increase timeout to 60 seconds as it can take monit up to 45 seconds to start
reason: On Bullseye Monit (5.27.2) the process will *sometimes* take the
configured start delay in time before it becomes available again. This
only seems to be a problem when it is restarted in quick succession.
Might be fixes in 5.31. In the meantime the timeout of 10 seconds was
insufficient in those cases and was increased to 60 seconds to allow for
the startup time.